Launch initramfs part as systemd service
Now that we're fscking the root partition from the initramfs,
the fsck task is running at the same time as the repartition script.

To prevent this, we need a bit more control, so convert the repartition
script to a systemd service unit and make it finish execution before
fsck is run.

[endlessm/eos-shell#2819]