initramfs diet
Add an endless dracut.conf.d entry that reduces the set of
modules included in the initramfs to the ones we use.

Similarly reduce the amount of fscks included, and fs drivers.

The initramfs drops from 22mb to 15mb as a result.

However, with fewer modules, dracut's cmdline and initqueue helpers
no longer run. This means that eos-repartition can be started before
the root device exists, and the root= shell parsing no longer happens.
Adjust eos-repartition dependencies to require the root device.

[endlessm/eos-shell#2390]