Add boot service to resize extra storage
If the system has the extra storage device, repartition and resize the
filesystem to fill the entire disk. This adds a new service instead of
building off of eos-repartition.service and eos-firstboot because the
timing is different. eos-repartition runs from the initrd since it
operates on /, but the extra storage can be resized from the real /.
Likewise, since we have access to the OS prior to the extra storage
being mounted, we can repartition and resize in one go by ordering
before the mount.

[endlessm/eos-shell#4414]