resize: Force fsck before resizing filesystem
resize2fs bails out if the filesystem has not been checked first. The
reason we don't have this problem for the root filesystem resizing is
because the repartitioning happens in the initrd and the service runs
after systemd-fsck-root.service.

Since the extra storage is being repartitioned and resized in one pass,
we can't rely on systemd to fsck it before resizing. Handle the fsck
ourselves.

[endlessm/eos-shell#4414]