Don't run extra storage units unless directories exist
If for any reason the /var/endless-extra and /endless paths don't exist,
then don't try to runt the extra storage mount and service units.
Furthermore, ensure that /endless is not a symlink. Without this, if
tmpfiles makes the /endless symlink, the overlayfs will be mounted there
in a way we don't want. Better to continue with the symlink for some
functionality.

[endlessm/eos-shell#4535]