extra-resize: Trigger from mount unit
Make sure that systemd doesn't try to mount /var/endless-extra until the
storage resizing script has completed. Furthermore, don't add the
resizing to local-fs.target so that it's only pulled in when the mount
has been enabled. This uses Wants rather than Requires so that there's a
chance the extra storage can be mounted as is despite any errors when
resizing.

[endlessm/eos-shell#4535]