Add dynamically activated systemd swap unit
Move our swap handling out of fstab and into a systemd unit
where we can control it better.

Now swap is purely optional but will be automatically enabled if/when
the appropriate device node appears.

In addition to automake getting it wrong, dh-systemd and systemd-204
also get escaping wrong when enabling this unit (systemd-215 might do it
better). Enable it manually in postinst.

[endlessm/eos-shell#4778]