Always enable /endless overlay mount
In order to make the app management the same on single disk and split
disk units, always mount the overlayfs at /endless. This also allows
ostree to set / as immutable when deploying. For single disk units, the
lower /var/endless-extra directory will simply be empty.

The Wants dependency on var-endless\x2dextra.mount has been removed so
that single disk units don't block on a non-existent device. For split
units, the mount will be enabled in another way while the After ensures
that the ordering remains correct.

For people upgrading on converted systems, the mount is skipped if
/endless is a symlink.

[endlessm/eos-shell#4654]