Add compat service for enabling extra mount
Since the /endless mount has been enabled on all units, it no longer
pulls in the /var/endless-extra mount as a dependency. In new images,
this mount will be enabled for split images. For an ostree upgrade,
split disk units will need to have the mount enabled at runtime.

Add a service and script that statically enables the mount on split disk
units. The script checks if this is an appropriate unit by looking at
the DMI information and seeing if the root disk is an eMMC device.

[endlessm/eos-shell#4654]