Use upstream 'overlay' filesystem
Ubuntu's Linux 4.1 (a requirement for eos 2.4.0) does not provide the
'overlayfs' filesystem we have been using before the upgrade, but a
compatibility layer implemented on top of the upstream 'overlay'
filesystem (endlessm/linux@a61b539). Since we're moving to a new
codebase in any case, let's just use the upstream version directly.

For both 'overlay' and 'overlayfs' filesystems there is a 'workdir'
mount option that needs to be passed to rw mounts.

[endlessm/eos-shell#4518]