repartition: don't log to kmsg
The new sfdisk writes a large amount of data to stdout (~1kb) all
in a single write() call. The kernel rejects it with -EINVAL and
that causes sfdisk to exit with error, even though it actually
did the repartitioning OK.

There's no good reason to log to the kernel logs directly,
all stdout and stderr goes to the systemd journal anyway.