eos-extra-resize: Don't check if disk is in use
By default, sfdisk does a BLKRRPART ioctl to check if a disk is in use,
before modifying it's partition table.

This causes udev to recreate partition nodes and do some probing,
meaning that by the time that sfdisk comes to do a second RRPART after
modifying the partition table, it can't, because udev is using the disk.

Skip the first RRPART ioctl to avoid this issue.

[endlessm/eos-shell#6211]