Update boot.mount to reflect the /boot bind-mount
Current EOS images does not have a separate boot partition anymore, and
/sysroot/boot is bind-mounted to /boot in the initrd as part of OSTree's
switchroot implementation. The newer systemd shipping in EOS 2.7
automatically unmounts block-device backed units when the backing block
device is not around. Without this change, this new systemd behavior has
a side-effect of unmounting /boot on systems without a separate boot
partition.

This change also provides the possibility of dropping the manual
bind-mount in the initrd, since it is only needed after the rootfs is
already switched.

https://phabricator.endlessm.com/T11157