repartition: detect root device from systemd
To move away from having the root filesystem always identifed by a fixed
label, update endless-repartition.sh to query sysroot.mount for the root
device.

That introduces a complication in the service file, we can no longer
order after a fixed device such as dev-disk-by\x2dlabel-ostree.device.

To make this unit run after the root device becomes available, but before
fscking it, we would have to likely modify systemd's generators so that
the availability of the root device triggers a new target, which
endless-repartition could be a part of.

But there's no strong requirement for us to run before fsck. We can
instead run after fsck, at which point the root device is guaranteed to
be ready.

https://phabricator.endlessm.com/T11392