zram: Don't start on early boot
There is currently an ordering conflict with this unit, as services are
ordered after sysinit.target by default, and we currently ordered it
before swap.target, which happens before sysinit.target.

If we wanted to keep it on early boot we would need to add
DefaultDependencies=false and manually pull the needed dependencies, but
since this is not necessary, lets simply order it in the same way as
regular services.

https://phabricator.endlessm.com/T11603