robustness: initial implementation
Add a dracut service that checks a boot-flags file to decide which root
to boot.

* If there's no flags file for this root, the service will just boot.
* If the "successful" flag is 1, the service will remove the flags file
and boot, since a previous boot attempt was successful we assume the
root is good.
* If the "successful" flag is 0 but "tries_left" is >0, it will try to
boot this root, decreasing "tries_left".
* Finally, if both "successful" and "tries_left" are 0, it will roll
back to the previous root swapping the bootloader files and setting
"/proc/cmdline" to the old root. It will also signal userspace by
creating a "rolled-back" file.

Initially, eos-updater sets this file with "successful=0" and
"tries_left=1" so one boot is attempted. If the user is able to log in,
a DBus service will mark the root as successful (see the following
commit).