Support booting from image file on exFAT
Add infrastructure needed to support booting from an Endless image file
found on an exFAT partition. This is primarily implemented inside the
new eos-image dracut module.

The endless.image.device parameter should be generated by the bootloader,
either specifying the PARTUUID= of the partition that contains the Endless
image.

The endless.image.path parameter provides the path to the image file on
that filesystem.

A new generator handles these parameters and hooks up the final block
device setup to the normal systemd initramfs fsck/mount infrastructure.

The block device setup script identifies which disk extents are occupied
by the image file and creates a linear device-mapper device that
corresponds. This is easier and much better performing than going through
exfat-fuse.

The repartitioning code now supports working with the resultant
device-mapper setup. We no longer get size and start info from sysfs,
because the sysfs paths to access dm info are a bit harder to deduce,
and the 'start' attribute is not present. We now use blockdev to get
the size, and we read the partition start from the sfdisk dump.

All that should cover the case for the regular booting of an Endless
image stored on a windows filesystem.

The endless.live_boot parameter implementation additionally tends to a
special case for the combined USB live+installer disk. In this case,
the image file used for live boot needs to be left untouched and pristine
so that the installer can work with it too. When this parameter is
specified, all the block devices are created in read-only mode, no
filesystem resizing is attempted, the OS updater is disabled, and
overlayfs is used for (throw-away) writable filesystem backing.

The overlay setup had to be split in 2. Most of them need to be created
quite late during real-root boot to avoid fighting with ostree-remount,
but /etc needs to be done before we switch root.

https://phabricator.endlessm.com/T12481