repartition: update generated units after changing filesystem UUID
The units generated in early boot use paths under /dev/disk/by-uuid/.
Now that we change the root fs UUID on first boot in the initramfs,
those generated units are no longer correct.

It's not possible to tweak the values in the generated units directly,
because you have to run daemon-reload after changing unit files, and
daemon-reload actually re-runs the generators.

However it's quite easy to use systemd's other methods to override unit
file values, which is what we do here.

We make the changes in the initramfs's /etc, which is thrown away when
the kernel switches root. This closely matches what happens to the
generated units too; they are stored in /run which is a mount shared with
the real root, but the generated units are wiped out a moment later when
systemd runs the generators from the real root.

https://phabricator.endlessm.com/T12317