Map the host partition to a dm device
Map the host partition which contains the endless.img as a new device
mapper device, so both the user and the flasher code can access files on
that partition.

Since endless.img is mapped as a different device, we map access to the
blocks occupied by endless.img to an error device, which means any I/O
on these blocks will always fail.

https://phabricator.endlessm.com/T13135