eos-enable-extra-upgrade: Run before local-fs.target
We want /var/endless-extra to be mounted before local-fs.target
completes. In order to ensure that happens on upgrades from old systems,
we need the eos-enable-extra-upgrade service to run early so that the
local-fs.target.wants link for var-endless\x2dextra.mount has been
created.

Rework the dependencies and ordering so that the service can run after
normal mounts are done but before var-endless\x2dextra.mount. With the
service part of local-fs.target, the script no longer needs to manually
start the mount unit. That seems to result in a deadlock.

https://phabricator.endlessm.com/T12747