Allow mounting endless-image-device rw
This should be safe now we have logic to guard the image file in
exfat-fuse and ntfs-3g. In the dual-boot case, this allows users to
delete files from their Windows drive to their heart's content! In the
combined USB case, this allows us to (potentially) save a debug log to
the image device.

https://phabricator.endlessm.com/T13136