eos-enable-extra-upgrade: Start mount unit without blocking
It appears that manually starting the mount unit from the upgrade
service is required. What I thought would happen (and I thought what I
tested) was that daemon-reload would cause systemd to see that the mount
unit was wanted for local-fs.target and block before starting
local-fs.target. That doesn't appear to be the case with the systemd in
2.6.

So, bring back the call to start the mount unit from the upgrade
service. However, due to the upgrade service having
`Before=var-endless\x2dextra.mount`, this would deadlock if the mount
call was done synchronously since the upgrade service would never
complete. Instead, call `systemctl start --no-block` so that the upgrade
service can exit and let systemd satisfy the ordering. This is what's
wanted, anyway. We don't need the upgrade service to block on the
mounting.

https://phabricator.endlessm.com/T13889