eos-repartition-mbr: leave expanding and swap to first boot
When this script was written, we didn't have support for MBR in
the initramfs repartition code, so it had to take care of enlarging
the partition and adding swap.

We've now readded MBR support to the repartition code, so we can
remove the code that is duplicated here. Instead, we just add the
special marker that will cause the repartition code to run on boot.

https://phabricator.endlessm.com/T13849