reparition: Don't leave stale boot entries behind
Changing the ESP's UUID breaks the Endless UEFI boot entry, since it
uses it as part of the loader path. To avoid stale boot entries to
accumulate on the machine, lets remove the Endless boot entry pointing
to the original ESP's UUID.

We could create a new boot entry here, but lets leave this job to
fallback.efi, so we do not need to keep the bootloader's path in sync.

https://phabricator.endlessm.com/T14430