repartition: don't delete recovery partition at end of disk
Asus will start shipping units in China with a recovery partition
added at the end of the disk before Endless first boot.

We can detect this by its GPT partition type UUID and then avoid
deleting it.

Since swap is no longer (always) the final partition we can't
rely on sfdisk making it fill the remaining space, so we need to
calculate and specify it's final size.

https://phabricator.endlessm.com/T14827