image-boot: Enable iso9660 booting
Reading the extents directly from the iso9660 and mapping them using a
linear device-mapper device unfortunately is not feasible when booting
the ISO image from CDROM. This is because the backing CDROM device file
(/dev/sr0) has 2048-byte sectors, while our image has been created
using 512-byte sectors.

This basically means that when the applications try to access the device
file /dev/mapper/endless-image (created mapping some sectors from
/dev/sr0), they think that the device has 2048-byte sectors instead of
512-byte sectors, getting confused. One notable example is kpartx that
is unable to read the partition table and refuses to create any
partition mappings.

To prevent this we create a loopback device for the endless image and
then create a device-mapper linear device corresponding to it.

Testing BIOS booting:

CDROM:
$ qemu-system-x86_64 -m 512 -machine q35,accel=kvm -cdrom image.iso

HDA:
$ qemu-system-x86_64 -m 512 -machine q35,accel=kvm -hda image.iso

Signed-off-by: Carlo Caione <carlo@endlessm.com>