Provide service to fix broken flatpak desktop files
For 3.0.10 and 3.0.11, the image builder was changed so that it
installed from the eos3.0 refs instead of the eos3 refs so that there
wasn't interference from eos3.1. To make it look like eos3 apps were
installed, the app checkout directories were changed to look that way.
Unfortunately, the branch is also embedded in the desktop file when
flatpak exports it, and this was not updated.

Provide a boot service that tries to find and fix these desktop files.
This will run only on upgrade.

https://phabricator.endlessm.com/T15005