image-boot-dm-setup: use zero, not error, for holes
If there happens to be a hole just before the end of the disk, `mount`
and friends try to read it while probing the filesystem type. This
fails, so the drive can't be mounted. (`mount -t ntfs-3g
/dev/mapper/endless-image-device /mnt` works fine -- it is not ntfs-3g
which is reading inside the holes.)

The 'zero' dm target returns zeros when read and discards all writes,
like /dev/zero.

https://phabricator.endlessm.com/T15818