Add eos-fix-home-dir-permissions
Starting with 3.1.6, we default to dir mode 700 for any
new user home directories, but previously we followed
the debian default of 755.

This service runs once after an upgrade from any version
prior to 3.1.6, to change the dir mode of any existing home
directories from 755 to 700.  Once run, a stamp file is
created to avoid running again, so that any further
tweaks by an admin to directory permissions will not be undone.

https://phabricator.endlessm.com/T16858