add eos-migrate-image-defaults script & service
A script to run once on upgrade which moves /var/eos-image-defaults to
/var/lib and ensures /var/lib/eos-image-defaults/settings (dconf file-db
file) exists. The dconf user profile in /etc/dconf is updated to point
to the new path, and if this makes it identical to the system-provided
file in /usr/share/dconf then it is removed. A compatibility symlink is
created in /var from the old path to the new one, because there is are
some flatpak apps which refer to the old path and can't be guaranteed to
be updated at the same time as the OS.

The job is triggered from systemd when /var/lib/eos-image-defaults/settings
does not exist. If there is an old /var/eos-image-defaults folder, a
/var/lib/.eos-migrate-image-defaults stamp file is made for the duration
of the migration, the existence of which implies the process was interrupted,
so also triggers the systemd job so that the migration is completed on reboot.