tests: add simple runner
The tests are not hooked up to `make check` because we'd need plumbing
to propagate the "skipped" state out to automake (which treats exit(77)
as skip). I'd rather not have them appear to PASS if they're not
actually run.

py.test produces much better output but I don't want to add a
build-dep on it given these tests are not even run by `make check`