debian/rules: remove extra list of .service files
The purpose of override_dh_systemd_enable here is to pass --no-enable
for eos-firewall-localonly.service but enable all other services. But
listing the .service files directly is error-prone -- it means that any
new .service files added to the package need to be added to this list
too. eos-add-flatpak-apps-repos.service and
eos-migrate-image-defaults.service were not included here when they were
added to the package, so did not run on converted systems.

When invoked without any unit files, dh_systemd_enable operates on all
units installed by the package which contain an [Install] section.
Debhelper provides a -X option which can be used to exclude items from
processing. This better expresses the intent of this rule -- enable all
services, except eos-firewall-localonly -- and fixes the two existing
bugs.

https://phabricator.endlessm.com/T15704