live-boot: overlay /var after it is mounted
In 90b1f12 we moved the overlay of /var earlier in the boot process.
That change was made because ostree-remount would mount an empty
tmpfs over /var if it was not writeable. This behaviour was changed in
https://github.com/endlessm/ostree/commit/05d0ee5cbecd1287b87d38e969862a5d8b1f2e58
https://github.com/endlessm/ostree/commit/30705889cb867c18cdb7fed8e55dc46477c069fa
and as a result, /var does not have the correct contents on
live boots any more. This appears to be because the real /var is
bind-mounted later in the boot process by a var.mount unit generated
from ostree-impl-system-generator, and has no effect if we have already
mounted something on /var.

The var.mount generated by ostree looks like this:

    [Unit]
    ...
    After=ostree-remount.service
    Before=local-fs.target

It's pulled in by:

    $ readlink /run/systemd/generator/local-fs.target.requires/var.mount
    ../var.mount

It previously ran in parallel to our
eos-live-boot-overlayfs-setup.service, since it has identical ordering
requirements.  So we need to force the latter to run *after* ostree's
var.mount.

The net result looks a bit confusing (wrapped for legibility):

    live@endless:~$ mount | grep var
    /dev/mapper/endless-image3 on /var type ext4
        (ro,relatime,data=ordered)
    eos-live-var on /var type overlay
        (rw,relatime,
         lowerdir=/var,
         upperdir=/run/eos-live/var,
         workdir=/run/eos-live/var-workdir)
    eos-live-var on /sysroot/ostree/deploy/eos/var type overlay
        (rw,relatime,
         lowerdir=/var,
         upperdir=/run/eos-live/var,
         workdir=/run/eos-live/var-workdir)

But it appears to work correctly!

https://phabricator.endlessm.com/T17622