Install com.endlessm.EknServices from eos-sdk
When adding the new eos-sdk repo after an upgrade, make sure the only
installed copy of EknServices is the one published there. (We assume
that there is an internet connection since this script runs when we just
completed an OS upgrade.)

We remove any orphaned refs belonging to the old
com.endlessm.EknServices, though normal users would not have these
anyway.

https://phabricator.endlessm.com/T17863