import code from eos-kalite to move EknServices deploy to eos-sdk repo
The new eos-flatpak-change-origin script uses an ingenious function from
our hero, Mario Sanchez Prada, which rewrites the deploy file of an
installed Flatpak so we can change its origin. We use it here so that
the EknServices Flatpak on your system appears to come from the new
eos-sdk repository - it will make use of the new apps.Platform once
it's been upgraded.

https://phabricator.endlessm.com/T17863