eos-link-host-mali-driver: Put GL driver in extension
This symlinks the host GL Mali driver in
/usr/lib/arm-linux-gnueabihf/mali400-egl-x11 into a place where Flatpak
will hopefully look for it and mount it as an extension. This makes the
separate org.freedesktop.Platform.GL.mali450-blah-blah extension
unnecessary, as we can simply use the host's driver.

Requires a lib -> . symlink in the Mali driver's directory, since Flatpak
will look for a lib/ directory inside the extension.

https://phabricator.endlessm.com/T18288