Explicitly setup /sysroot/boot -> /boot bind mount
On OSTree systems, this is manually setup in the initrd by
ostree-prepare-root. But since it is not configured anywhere in the
system, systemd-gpt-auto-generator is not aware of it, and ends up
un-mounting it due to its own boot.mount and boot.automount units,
generated to mount the ESP on /boot.

We don't need to mount the ESP on EOS, and making this bind mount
explicit avoids this systemd-gpt-auto-generator's behaviour.

This effectively reverts commit
ccb3e26770ebbc7709a689497e7fbd59175e2638.

https://phabricator.endlessm.com/T18825