Activate nvidia graphics driver on supported hardware
As nouveau is not supporting the latest nvidia hardware well enough,
we will now activate the nvidia driver automatically, when we detect
supported hardware.

Detection is done using PCI IDs, using the same technique
as Debian.

We have to remove nouveau from the initramfs, and blacklist it on
the real root, in order to stop it being loaded, since otherwise
we cannot unload it (to then load nvidia).

However, if nvidia load then fails, we fall back to nouveau.
We also fall back on nouveau if the user has blacklisted
nvidia in the regular modprobe configuration.

https://phabricator.endlessm.com/T19190