Add service to prevent adding ostree remotes in /etc/ostree/remotes.d
Set the repo config option core.add-remotes-config-dir repo option to
false so that new remotes are not added in /etc/ostree/remotes.d. Adding
them there is incompatible with the way flatpak modifies remotes and
will cause the repo to become unusable. Setting that option to false
ensures that ostree only adds remotes in the repo's config file.

This might only needed temporarily until ostree and flatpak get updated
to do sane remote modification.

The service is only run on updates and before any known /ostree/repo
services. The other services here that modify the repo are setup to run
after this service. That was done to increase the likelihood that new
services copied from those will also run after this service.

https://phabricator.endlessm.com/T19077
https://github.com/ostreedev/ostree/issues/1134