eos-link-host-mali-driver: Fix host path
With the work to support NVIDIA binary drivers we had to move our Mali
drivers to a different location. We need to adjust paths here
accordingly, overwriting the old link in case it exists.

ConditionPathExists= follows symbolic links, so the service will run
when upgrading on a system which already has the system pointing to the
old location.

https://phabricator.endlessm.com/T18579