nvidia-graphics: remove modules after load on first boot
We don't want to distribute the nvidia modules from the factory-booted
system.

The initial attempt to do this with nvidia-graphics-cleanup
does not work when the operator has entered Test Mode, which creates
overlayfs mounts which are still in place when nvidia-graphics-cleanup
runs. At this point it appears that user accounts are present and
we would not be able to remove the nvidia drivers from disk anyway.

Solve this by deleting the nvidia drivers right after they are loaded,
if no user accounts are present. This means that when the system reaches
the user they will link the modules both during first and second boot,
but it is so quick we don't expect anyone to notice.

https://phabricator.endlessm.com/T18676