Combine flatpak configuration into a single service
While troubleshooting a split-disk EC-200 that was upgraded
from EOS 2.6.0 to EOS 3.3.0 (staging), some of these services
failed, apparently due to an interprocess race that may overwrite
the repo/config file in incompatible ways.

For now, the individual scripts remain separate and are just
called by one master script.  Down the road it might be nice
to combine them into a single Python script, but that would
require some refactoring since some are written in Bash.

https://phabricator.endlessm.com/T19491