Setup flatpak extension directory on extra storage
On non-split systems /var/endless-extra does not exist, so if try to
create this symlink through a static tmpfile, systemd-tmpfiles returns
an error (although the other files are created correctly), causing
systemd-tmpfiles-setup.service to exit with a failure status.

Let's use a simple service unit to create it only on split-disk systems.

https://phabricator.endlessm.com/T19691