Re-enable gnome runtime repo
Some of our Endless-built apps depend on the gnome-3.22 and
freedesktop-1.4 runtimes, which are not available in Flathub.

We still keep the removal of the gnome-apps repo, which is
what actually causes the app duplication in the app center.

We can revert this once we have all of our apps migrated.

This is a partial revert of fd15723e318dea1082ff5deb64a1ffeda3ee6c41

https://phabricator.endlessm.com/T19898