Don't remove gnome-apps repo
Until we have a mechanism to transition installed gnome-apps apps
to repoint to flathub, users will be left without the ability
to update apps that they have already installed from gnome-apps.

We still don't install gnome-apps for new users, we merely
do not remove it from existing users.

We can revert this once we have a migration plan in place.

This is a partial revert of fd15723e318dea1082ff5deb64a1ffeda3ee6c41

https://phabricator.endlessm.com/T19898