eos-repartition-mbr: preserve ESP
Previously, we removed it from the partition table. But in practice it
is relatively common for users (and Endless developers!) to boot from
USB stick in legacy mode, which causes eos-installer to run this script
after writing the image to disk, then attempt to boot the installed
system in UEFI mode.

Instead, we preserve the ESP, using the 'EF' partition type.  This is
the same partition type we and others use in our MBR-only hybrid .iso
images, so it's expected that the majority of UEFI systems will be able
to boot it.  (eos-installer still prefers to leave the disk in GPT
format if it know the system is UEFI.)

While we're here, let's remove the last vestige of swap partition
handling, which should have been removed as part of
9fdbc5ee1f2abd64fb4915422a18c9104091ec47. This script should always be
run against a freshly-written, pristine image.

https://phabricator.endlessm.com/T20245