eos-repartition-mbr: don't reinstall bootloader
Previously, we would re-install the bootloader to the boot track. This
worked okay, but there's no pressing reason to do it when the space
previously allocated to the BIOS boot partition is still there; leaving
it out makes the script simpler, and incidentally easier to test.

I think my rationale previously was that a user might tinker with their
partition table and overwrite the bootloader. I guess that's possible
but it's definitely in "this voids your warranty" territory. GNOME
Disks, at least, does not display the 1MiB of unpartitioned space
between the ESP and ostree partitions at all.

https://phabricator.endlessm.com/T20245