migrate Steam from eos-apps to flathub
Introduces a new function into eos-add-flatpak-apps-repos which allows us to
provide an array of flatpak names (either apps or runtimes) which are matched
by branch and/or origin, and moved atomically to the new branch and/or origin
as requested.

Changing the origin is achieved with our existing code for rewriting deploy
files, and changing the branch is done by hardlinking the old deploy folder
to the new branch name, and updating the "current" symlink if it's an app
(runtimes have no current symlink as they are always referenced using the
branch name).

This means we can move Steam from the eos-apps origin to flathub, and update
the branch from eos3 to stable. The user can then upgrade their Steam and get
working 32-bit NVidia drivers from upstream Flathub.

https://phabricator.endlessm.com/T20322