migrate branch in .desktop and .service files
When modifying the branch name in /var/lib/flatpak, we must also
update the exported files which Flatpak patches during deploy time
to invoke flatpak run --branch=... with the new branch name. This
updates .desktop and .service files so that they launch correctly.
These files are removed and re-created the next time the flatpak
is upgraded, but there is no API to re-deploy without pulling.