nvidia-graphics-setup: enable HDMI audio device
Some BIOSes (like the Asus GL502VS) disable the HDMI audio device
during boot. Nvidia showed how this can be controlled by a bit in
PCI extended configuration space.

Under Windows, this bit appears to be managed according to whether a
HDMI cable is connected or not.

Until this issue is better solved in the driver, attempt to detect
this case and apply the workaround before loading the driver.

https://devtalk.nvidia.com/default/topic/1024022/linux/gtx-1060-no-audio-over-hdmi-only-hda-intel-detected-azalia
https://bugs.freedesktop.org/show_bug.cgi?id=75985
https://phabricator.endlessm.com/T20286