Add eos-fix-external-appstream
EOS downloads an external appstream file "eos-extra.xml.gz" every time
there's a refresh and the server's file is younger than the cached one.

We ship a version of eos-extra.xml.gz in the images as a base point but
the name used to store the file no longer corresponds to the one GNOME
Software generates: it should now take an "org.gnome.Software-" prefix.

So this patch creates a service to rename or remove that file.

https://phabricator.endlessm.com/T19319